+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # PineappleTimer
 
 ## [Unreleased]
+- Add 5 minutes rest notification
 
 ## [watchOS 1.0.0(3)] - 2019-09-18
 - 解決 complication 沒有更新的問題